+++ Changes.md
@@ -67,6 +67,28 @@ Build
   - OpenColorIO : Updated to version 2.1.1.
   - Cortex : Updated to version 10.4.0.0.
 
+0.61.x.x (relative to 0.61.4.0)
+========
+
+Fixes
+-----
+
+- Spreadsheet : Fixed errors computing `resolvedRows` when the Spreadsheet contains a NameValuePlug (#4583).
+- Viewer : Fixed points drawing mode for CurvesPrimitives.
+
+API
+---
+
+- Tool : Added support for subclassing in Python.
+- View : Added `tools()` method which returns a container of all connected Tools.
+- Container : Added constructor with a `name` argument.
+- PlugAlgo : Added support for NameValuePlug in `extractDataFromPlug()`.
+
+Build
+-----
+
+- Cortex : Updated to version 10.3.3.0.
+
 0.61.4.0 (relative to 0.61.3.0)
 ========